refactor: extract CertificateDrawer Let's Encrypt and custom cert sections
Part of #103
