+++ src/components/features/certificates/CertificateDrawer.tsx
@@ -1,33 +1,22 @@
 import React from 'react'
 import {
   TextField,
-  FormControlLabel,
-  Switch,
   Box,
   Alert,
-  Button,
-  Stack,
-  CircularProgress,
 } from '@mui/material'
 import {
   Info as InfoIcon,
-  CheckCircle as CheckIcon,
-  Description as FileIcon,
-  Key as KeyIcon,
-  AccountTree as ChainIcon,
 } from '@mui/icons-material'
 import { Certificate, CreateCertificate, certificatesApi } from '../../../api/certificates'
-import { getErrorMessage } from '../../../types/common'
 import BaseDrawer from '../../base/BaseDrawer'
 import FormSection from '../../shared/FormSection'
-import DomainInput from '../../DomainInput'
 import { useDrawerForm } from '../../../hooks/useDrawerForm'
-import FileDropzone from './components/FileDropzone'
-import DNSProviderSelector from './components/DNSProviderSelector'
 import { useToast } from '../../../contexts/ToastContext'
 import { NAVIGATION_CONFIG } from '../../../constants/navigation'
 import { TIMING } from '../../../constants/timing'
 import { LAYOUT } from '../../../constants/layout'
+import LetsEncryptSection from './LetsEncryptSection'
+import CustomCertificateSection from './CustomCertificateSection'
 
 interface CertificateDrawerProps {
   open: boolean
@@ -37,7 +26,7 @@ interface CertificateDrawerProps {
   initialProvider?: 'letsencrypt' | 'other'
 }
 
-interface CertificateFormData {
+export interface CertificateFormData {
   provider: 'letsencrypt' | 'other'
   niceName: string
   domainNames: string[]
@@ -62,8 +51,6 @@ export default function CertificateDrawer({
 
   const { showSuccess, showError } = useToast()
 
-  const [testingDomains, setTestingDomains] = React.useState(false)
-  const [testResult, setTestResult] = React.useState<{ reachable: boolean; error?: string } | null>(null)
   const [customError, setCustomError] = React.useState<string | null>(null)
 
   const isEditMode = !!certificate
@@ -317,15 +304,6 @@ export default function CertificateDrawer({
     }
   })
 
-  // Memoize callbacks to prevent unnecessary re-renders in DNSProviderSelector
-  const handleDnsProviderChange = React.useCallback((provider: string) => {
-    setFieldValue('dnsProvider', provider)
-  }, [setFieldValue])
-  
-  const handleDnsCredentialsChange = React.useCallback((credentials: string) => {
-    setFieldValue('dnsProviderCredentials', credentials)
-  }, [setFieldValue])
-
   // Update form data when certificate prop changes (for edit mode)
   React.useEffect(() => {
     if (certificate && open) {
@@ -362,27 +340,6 @@ export default function CertificateDrawer({
     }
   }, [open])
 
-  const handleTestDomains = async () => {
-    if (!data || data.domainNames.length === 0) {
-      return
-    }
-    
-    setTestingDomains(true)
-    setTestResult(null)
-    
-    try {
-      const result = await certificatesApi.testHttpReachability(data.domainNames)
-      setTestResult(result)
-    } catch (err: unknown) {
-      setTestResult({ 
-        reachable: false, 
-        error: getErrorMessage(err)
-      })
-    } finally {
-      setTestingDomains(false)
-    }
-  }
-
   // Determine the current provider - use data if available, otherwise use initial values
   const currentProvider = data?.provider || certificate?.provider || initialProvider
 
@@ -409,170 +366,25 @@ export default function CertificateDrawer({
           </Alert>
         )}
 
-        {/* Domain Names Section - Only for Let's Encrypt */}
+        {/* Let's Encrypt form (domains + LE config + challenge method) */}
         {currentProvider === 'letsencrypt' && (
-          <FormSection title="Domain Names" required>
-            <DomainInput
-              value={data.domainNames}
-              onChange={(domainNames) => setFieldValue('domainNames', domainNames)}
-              helperText="Enter the domain names this certificate should cover. Wildcards (*.example.com) are supported."
-              error={Boolean(errors.domainNames && touched.domainNames)}
-              required
-              disabled={isEditMode}
-            />
-            {errors.domainNames && touched.domainNames && (
-              <Alert severity="error" sx={{ mt: 1 }}>{errors.domainNames}</Alert>
-            )}
-          </FormSection>
+          <LetsEncryptSection
+            data={data}
+            errors={errors}
+            touched={touched}
+            setFieldValue={setFieldValue}
+            isEditMode={isEditMode}
+          />
         )}
 
-
-        {/* Let's Encrypt Configuration */}
-        {currentProvider === 'letsencrypt' && !isEditMode && (
-          <>
-            <FormSection title="Let's Encrypt Configuration" required>
-              <TextField
-                label="Email Address"
-                type="email"
-                value={data.letsencryptEmail}
-                onChange={(e) => setFieldValue('letsencryptEmail', e.target.value)}
-                helperText="Used for important certificate notifications"
-                required
-                fullWidth
-                sx={{ mb: 2 }}
-              />
-
-              <FormControlLabel
-                control={
-                  <Switch
-                    checked={data.letsencryptAgree}
-                    onChange={(e) => setFieldValue('letsencryptAgree', e.target.checked)}
-                  />
-                }
-                label="I agree to the Let&apos;s Encrypt Subscriber Agreement"
-              />
-            </FormSection>
-
-            <FormSection title="Challenge Method">
-              <FormControlLabel
-                control={
-                  <Switch
-                    checked={data.dnsChallenge}
-                    onChange={(e) => setFieldValue('dnsChallenge', e.target.checked)}
-                  />
-                }
-                label="Use DNS Challenge"
-              />
-              
-              {data.dnsChallenge ? (
-                <>
-                  <Alert severity="info" sx={{ mt: 1 }}>
-                    DNS challenge allows wildcard certificates and works behind firewalls.
-                  </Alert>
-                  
-                  <Box sx={{ mt: 2 }}>
-                    <DNSProviderSelector
-                      value={data.dnsProvider}
-                      onChange={handleDnsProviderChange}
-                      credentials={data.dnsProviderCredentials}
-                      onCredentialsChange={handleDnsCredentialsChange}
-                    />
-
-                    <TextField
-                      label="Propagation Seconds"
-                      type="number"
-                      value={data.propagationSeconds}
-                      onChange={(e) => setFieldValue('propagationSeconds', parseInt(e.target.value) || 120)}
-                      helperText="Time to wait for DNS propagation (default: 120 seconds)"
-                      fullWidth
-                      sx={{ mt: 2 }}
-                    />
-                  </Box>
-                </>
-              ) : (
-                <>
-                  <Alert severity="info" sx={{ mt: 1 }}>
-                    HTTP challenge requires domains to be publicly accessible on port 80.
-                    Wildcard certificates are not supported with HTTP challenge.
-                  </Alert>
-                  
-                  <Box sx={{ mt: 2 }}>
-                    <Button
-                      variant="outlined"
-                      onClick={handleTestDomains}
-                      disabled={testingDomains || data.domainNames.length === 0}
-                      startIcon={testingDomains ? <CircularProgress size={20} /> : <CheckIcon />}
-                    >
-                      Test Domain Reachability
-                    </Button>
-                    {testResult && (
-                      <Alert
-                        severity={testResult.reachable ? 'success' : 'error'}
-                        sx={{ mt: 2 }}
-                      >
-                        {testResult.reachable
-                          ? 'All domains are reachable and ready for certificate generation'
-                          : testResult.error || 'Some domains are not reachable'}
-                      </Alert>
-                    )}
-                  </Box>
-                </>
-              )}
-            </FormSection>
-          </>
-        )}
-
-        {/* Custom Certificate Upload */}
+        {/* Custom certificate upload form */}
         {currentProvider === 'other' && !isEditMode && (
-          <>
-            <FormSection title="Certificate Name" required>
-              <TextField
-                label="Certificate Name"
-                value={data.niceName}
-                onChange={(e) => setFieldValue('niceName', e.target.value)}
-                placeholder="e.g. Production SSL Certificate"
-                fullWidth
-                required
-                error={Boolean(errors.niceName && touched.niceName)}
-                helperText={errors.niceName && touched.niceName ? errors.niceName : "A friendly name to identify this certificate"}
-              />
-            </FormSection>
-
-            <FormSection title="Certificate Files" required>
-              <Stack spacing={3}>
-              <FileDropzone
-                label="Certificate File"
-                icon={<FileIcon color="action" />}
-                file={data.certificateFile}
-                onFileSelect={(file) => setFieldValue('certificateFile', file)}
-                accept=".pem,.crt,.cer"
-                required
-                validateType="certificate"
-                helperText="The SSL certificate file (should start with -----BEGIN CERTIFICATE-----)"
-              />
-
-              <FileDropzone
-                label="Private Key"
-                icon={<KeyIcon color="action" />}
-                file={data.certificateKeyFile}
-                onFileSelect={(file) => setFieldValue('certificateKeyFile', file)}
-                accept=".key,.pem"
-                required
-                validateType="key"
-                helperText="The private key file (should start with -----BEGIN PRIVATE KEY-----)"
-              />
-
-              <FileDropzone
-                label="Intermediate Certificate (optional)"
-                icon={<ChainIcon color="action" />}
-                file={data.intermediateCertificateFile}
-                onFileSelect={(file) => setFieldValue('intermediateCertificateFile', file)}
-                accept=".pem,.crt,.cer"
-                helperText="Optional intermediate certificate for certificate chain validation"
-              />
-              </Stack>
-            </FormSection>
-          </>
+          <CustomCertificateSection
+            data={data}
+            errors={errors}
+            touched={touched}
+            setFieldValue={setFieldValue}
+          />
         )}
 
         {/* Nice Name - Show for Let's Encrypt (optional) and in edit mode */}
